Added basic gui text implementation

+++ src/main/java/keithcod/es/fusionengine/client/engine/IGameLogic.java
@@ -6,7 +6,7 @@ public interface IGameLogic {
 
     void input(Window window, Input input);
 
-    void update(float interval, Input input);
+    void update(double interval, Input input);
 
     void render(Window window);
